+++ pkg/transform/transform.go
@@ -25,6 +25,15 @@ func isMDFile(path string) bool {
 	return filepath.Ext(path) == ".md"
 }
 
+func isImgFile(path string) bool {
+	switch filepath.Ext(path) {
+	case ".png", ".jpg", "jpeg", ".svg", ".gif", ".webp":
+		return true
+	default:
+	}
+	return false
+}
+
 func prepOutputDir(d string, gitIgnored bool) error {
 	_, err := os.Stat(d)
 	if err != nil && !os.IsNotExist(err) {
@@ -65,11 +74,12 @@ func Dir(ctx context.Context, logger log.Logger, config []byte) error {
 		logger: logger,
 
 		linkTransformer: &relLinkTransformer{
-			localLinksStyle: c.LocalLinksStyle,
-			inputDir:        c.InputDir,
-			outputDir:       c.OutputDir,
-			oldRelPath:      map[string]string{},
-			newRelPath:      map[string]string{},
+			localLinksStyle:                   c.LocalLinksStyle,
+			inputDir:                          c.InputDir,
+			outputDir:                         c.OutputDir,
+			oldRelPath:                        map[string]string{},
+			newRelPath:                        map[string]string{},
+			linkPrefixForNonMarkdownResources: c.LinkPrefixForNonMarkdownResources,
 		},
 	}
 
@@ -259,10 +269,11 @@ func (t *transformer) transformFile(path string, info os.FileInfo, err error) er
 type relLinkTransformer struct {
 	localLinksStyle LocalLinksStyle
 
-	inputDir   string
-	outputDir  string
-	oldRelPath map[string]string
-	newRelPath map[string]string
+	inputDir                          string
+	outputDir                         string
+	oldRelPath                        map[string]string
+	newRelPath                        map[string]string
+	linkPrefixForNonMarkdownResources string
 }
 
 func (r *relLinkTransformer) TransformDestination(ctx mdformatter.SourceContext, destination []byte) ([]byte, error) {
@@ -294,6 +305,30 @@ func (r *relLinkTransformer) TransformDestination(ctx mdformatter.SourceContext,
 		if err != nil {
 			return nil, errors.Wrap(err, "link: clean old dest path")
 		}
+	}
+
+	// Non md or image relative link, so needs link to be prefixed.
+	if r.linkPrefixForNonMarkdownResources != "" && !isMDFile(relDest) && !isImgFile(relDest) {
+		workingDir, err := os.Getwd()
+		if err != nil {
+			return nil, err
+		}
+		// Original path before transform needs to be figured out.
+		// Case where file is in same position after transform.
+		oldRelPath := filepath.Join(r.inputDir, curRelDir, relDest)
+		if _, err := os.Stat(oldRelPath); os.IsNotExist(err) {
+			// File is moved to new dir after transform.
+			oldRelPath = filepath.Join(workingDir, curRelDir, relDest)
+			if _, err := os.Stat(oldRelPath); os.IsNotExist(err) {
+				// File is extra input glob.
+				oldRelPath = filepath.Join(workingDir, relDest)
+			}
+		}
+		originalRelPath, err := filepath.Rel(workingDir, oldRelPath)
+		if err != nil {
+			return nil, err
+		}
+		return []byte(r.linkPrefixForNonMarkdownResources + "/" + originalRelPath), nil
 	}
 
 	currDest := oldRelDest